(#138) ui_inventory basics !
ui_inventory shows inventory's items, scalable or not, and update when the inventory grab/drop one.

for now the ui_inventory is not linked to any OpenCapacity, and we can't grab/drop items from it.

updated the ItemBank a little bit to help with the ui_inventory. also the Item now have a string Reference which is basically the name of the item. useful for naming the items in the hierarchy differently.