(#153) first pathfinding action working !
used the A* project to make a pathfinding grid and made a Chase.cs Goal using a GoToAstar.cs Action that implements the pathfinding to control a zombo.

BUT i watched a great video on GOAP system SO i will use the free asset GOAP v3 from the asset store which is a banger and allow me not to spend 1 year building a GOAP system that will not even be half-good compared to this asset.

good news !! so next commit is the removal of Goal.cs, Action.cs and all that shit bye bye oopsie

rework the fridge skin and added fridge.hover.D