(#154.6) Switched to 2 Sensors instead of a MultiSensor for EatGoal + fixed bugs
moved the hunger variable into IA and made an FoodRule string variable to filter food that the ia wants to eat. This rule use the new Item.ValidateRule(string rule) method to easily check those bools

made a new zombo agent type with its factory BUT had an error when getting the ia into MultiSensor.Update() method to apply the FoodRule but that was working great with only cats... i think Zombo brain was not requesting for the EatGoal but had the AddCapability<EatCapabilityFactory>() into its factory maybe that was why?

So i switched to 2 sensors and got the ia ref into the ClosestFoodSensor.Sense() method

fixed a bug into the EatAction that was eating the food even if the food was not Eatable anymore when BeforePerforming the action.

performance tests, zombo + cats eating all those things together, got up to ~200 mobs together !!!!!