(#167.4) UI_InventoryMenu disabling drop/activate/move action when not useful
UI_Navigator has now a OnSlotHoverEnter event that gives the I_UI_Slot just hovered.
Also has a new ToggleInput(string, bool) method to enable/disable precise callback at runtime

UI_InventoryMenu register to this event and checks if we need to disable/enable the callbacks
It also calls the InputFeedback's right methods to update the IF accordingly

InputFeedback now has an optional reference to a Label (TMPro.idkthelettersText_UI_GUI) and a SetLabel(string) method to change it.
ButtonFeedback has also a SetAlwaysFull(bool) method to change this

it's working great !