(#224) UI_Oven !! + UI_OnOffToggle omgg so satisfiyng
cool ui :D

i thought about re re re re changing the in-game ui for the player to be able to move inside ui with Ljoy instead of Rjoy
BUT we stick to the B button for taking items inside the chest. SO we need to put a UI_Button with a cross (RU in windows) that will close the chest when B pressed on it. better so we can move through the inventory with Ljoy and, if well done, the ui_navigator will automatically click on it if we hold B for taking all objects -> niceee