(#231) UI_Navigator rework part 1 !!!
omg lot of stuff

added UI_Navigator that will replace UI_XboxNavigator.
Moved most of navigator inputs to UIC (all except move item for now)
UI_Navigator works with directly a UI_Slot as the current slot and not GameObject's index as it was previously the case (was dreadful)
UI_Navigator also works with new UI_Slottable :

added a new UI_Slottable which will replace I_UI_Slottable. UI_Slottable is now the mother of UI_Inventory, UI_Slotter, UI_ParentBasedSlottable & UI_SlottableMixer that all handle slots and has a saved slot to return to.
when enabling an UI_Slottable, it goes directly back to its saved position. UI_SlottableMixer is a "mixer" as you can tell from the name, so it does not have slots but has a List<UI_Slottable> and a UI_Slottable master. it allows me to remove the Slottable logic on UI_Pool by putting a UI_SlottableMixer on them that will have all pool's specific slottables. and so we only need to enable the ui_navigator for 1 single UI_Slottable (the mixer) and all derivative slottables will be enabled. the master on this mixer may be assigned to get back to the preferred slottable's starting slot when enabling the mixer. (ex : when we open a chest, we want to hover the chest's first slot and not the perso quick inventory -> here the master is the ui_slottable of the chest. BUT in the inventory menu, we don't want to choose between ui_inventory & ui_laptop -> we don't have a master, leave it to null)

removed dependency to I_UI_Slot, now UI_Slot is the mother of all slots, and added a new UI_ImageSlot which is the mother of UI_Item, UI_Module, UI_Button, UI_Toggle. and now UI_Skill & UI_Text are children of UI_Slot directly, which is greaaat

few bugs found :
- stacking a pool on the hud won't disable the stacked pools, which means if we stack a slottable pool on an open ui_chest, it will break things bcz both slottables will be enabled